dcp-t4(003): move diagnostics and workerinference into factory_definitions
Relocate pkg/transports/mapping/factoryconfig/diagnostics and
pkg/transports/mapping/workerinference to
pkg/services/factory_definitions/transports/mapping/, preserving the mapping
protocol segment and the factoryconfig/ segment above diagnostics. Both map
Factory Definitions state and import only that service among product services.

Pure relocation: file contents are unchanged apart from import paths.

diagnostics moved with no consumer change. Its two importers are both tests,
and cmd/pkgboundarycheck does not raise the peer-service subpackage rule on
_test.go files, so pkg/transports/http/servertests keeps importing it directly.

workerinference's two importers are production files in a peer service,
pkg/services/models/transports/{http,cli}. cmd/pkgboundarycheck rejects a peer
service reaching into another service's transport subtree: main.go:1871
isMatchingServiceOwnedTransportConsumer only excepts consumers under the
matching composition-root protocol root, which pkg/services/models is not.

Both now consume pkg/transports/mapping/factorydefinitionentry, the
composition-root entry package #2060 added for this exact case; its package doc
already states that peer-service transports consume it "instead of reaching
into that service's transport subpackages directly". Only the single function
those callers use, OperationBindingsFromGenerated, is republished, and it
delegates to the service-owned implementation, so no logic is duplicated.

The import alias is dropped rather than kept as workerinferencemapping, so the
call sites name the package they actually import.

cmd/servicecyclecheck still measures minimum feedback arc weight 22, matching
the ceiling, and this diff adds no package-boundary violation.
