+++ pkg/services/factory_definitions/transports/cli/factoryload/blocking_factory_load_test.go
@@ -8,7 +8,7 @@ import (
 
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/transports/cli/factoryload"
-	configdiagnostics "github.com/portpowered/infinite-you/pkg/transports/mapping/factoryconfig/diagnostics"
+	configdiagnostics "github.com/portpowered/infinite-you/pkg/services/factory_definitions/transports/mapping/factoryconfig/diagnostics"
 )
 
 func TestBlockingFactoryLoadFindings_PreservesCanonicalTargets(t *testing.T) {
