+++ pkg/services/factory_definitions/transports/mapping/factoryconfig/diagnostics/diagnostics_test.go
