+++ pkg/services/factory_definitions/transports/mapping/workerinference/generated.go
