+++ pkg/services/models/transports/cli/bootstrap_invoke.go
@@ -12,8 +12,8 @@ import (
 	modelinference "github.com/portpowered/infinite-you/pkg/services/models"
 	operatorconfig "github.com/portpowered/infinite-you/pkg/services/operator_settings"
 	"github.com/portpowered/infinite-you/pkg/services/work"
+	"github.com/portpowered/infinite-you/pkg/transports/mapping/factorydefinitionentry"
 	contentcontract "github.com/portpowered/infinite-you/pkg/transports/mapping/workcontent"
-	workerinferencemapping "github.com/portpowered/infinite-you/pkg/transports/mapping/workerinference"
 	"go.uber.org/zap"
 )
 
@@ -205,7 +205,7 @@ func invocationRequestFromGenerated(request factoryapi.ModelInvocationRequest) m
 }
 
 func modelBindingsFromGenerated(bindings *[]factoryapi.WorkstationOperationBinding) []modelinference.ModelOperationBinding {
-	authored := workerinferencemapping.OperationBindingsFromGenerated(bindings)
+	authored := factorydefinitionentry.OperationBindingsFromGenerated(bindings)
 	result := make([]modelinference.ModelOperationBinding, len(authored))
 	for i := range authored {
 		result[i] = modelinference.ModelOperationBinding{
